Fixed bug with unsigned to f64 conversions
Conversion of unsigned/signed integer of various sizes to 64-bit floating point number did not work correctly. Now they do.

I have not tested f64 to integers and that probably does not work. (fix it next time?)

+++ wa.cpp
@@ -7,7 +7,7 @@ int main() {
     // auto a = hey();
 
     unsigned long int j = 235;
-    float k = j;
+    double k = j;
     
     return 0;    
 }
